enforce runtime as nodejs. Make new capture form server component page for robust session user check'
