+++ src/app/(signed-in)/capture/new/components/app-gallery.tsx
