+++ src/app/api/auth/[...nextauth]/route.ts
@@ -1,2 +1,6 @@
 import { handlers } from "@/lib/auth/auth";
+
+// Force Node.js runtime to avoid edge compatibility issues
+export const runtime = "nodejs";
+
 export const { GET, POST } = handlers;
